refactor: organize DTO imports and update package structure
- Moved `ResetPasswordRequest` to `dto.request` package.
- Removed unused imports across several classes (`AuthService`, `AuthController`, `AuthServiceImpl`, etc.).
- Optimized imports for better readability and package organization.
